Move `cmd_ctx_builder` support code into `test_support.rs`.
