+++ workspace_tests/src/cmd/ctx/cmd_ctx_builder/single_profile_no_flow_builder.rs
@@ -4,7 +4,7 @@ use peace::{
     resources::paths::{ProfileDir, ProfileHistoryDir},
 };
 
-use crate::{cmd::ctx::cmd_ctx_builder::workspace, no_op_output::NoOpOutput, PeaceTestError};
+use crate::{no_op_output::NoOpOutput, test_support::workspace, PeaceTestError};
 
 #[tokio::test]
 async fn build() -> Result<(), Box<dyn std::error::Error>> {
