+++ workspace_tests/src/lib.rs
@@ -37,4 +37,5 @@ mod fn_tracker_output;
 mod fn_tracker_presenter;
 mod no_op_output;
 mod peace_test_error;
+mod test_support;
 mod vec_copy_item;
